+++ marathon-be/src/modules/cms/cms.controller.js
@@ -1,3 +1,4 @@
+import { validationResult } from "express-validator";
 import { cmsService } from "./cms.service.js";
 import { successResponse, errorResponse } from "../../utils/response.js";
 import { asyncHandler } from "../../utils/asyncHandler.js";
@@ -8,6 +9,14 @@ export const listPages = asyncHandler(async (req, res) => {
 });
 
 export const getPage = asyncHandler(async (req, res) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return errorResponse(res, {
+      statusCode: 400,
+      message: errors.array()[0].msg,
+    });
+  }
+
   const result = await cmsService.getPage(req.params.id);
   return successResponse(res, { data: result });
 });
@@ -18,16 +27,40 @@ export const getPageBySlug = asyncHandler(async (req, res) => {
 });
 
 export const createPage = asyncHandler(async (req, res) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return errorResponse(res, {
+      statusCode: 400,
+      message: errors.array()[0].msg,
+    });
+  }
+
   const result = await cmsService.createPage(req.body);
   return successResponse(res, { statusCode: 201, data: result });
 });
 
 export const updatePage = asyncHandler(async (req, res) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return errorResponse(res, {
+      statusCode: 400,
+      message: errors.array()[0].msg,
+    });
+  }
+
   const result = await cmsService.updatePage(req.params.id, req.body);
   return successResponse(res, { data: result });
 });
 
 export const deletePage = asyncHandler(async (req, res) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return errorResponse(res, {
+      statusCode: 400,
+      message: errors.array()[0].msg,
+    });
+  }
+
   await cmsService.deletePage(req.params.id);
   return successResponse(res, { message: "Page deleted" });
 });
@@ -38,16 +71,40 @@ export const listAnnouncements = asyncHandler(async (req, res) => {
 });
 
 export const createAnnouncement = asyncHandler(async (req, res) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return errorResponse(res, {
+      statusCode: 400,
+      message: errors.array()[0].msg,
+    });
+  }
+
   const result = await cmsService.createAnnouncement(req.body);
   return successResponse(res, { statusCode: 201, data: result });
 });
 
 export const updateAnnouncement = asyncHandler(async (req, res) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return errorResponse(res, {
+      statusCode: 400,
+      message: errors.array()[0].msg,
+    });
+  }
+
   const result = await cmsService.updateAnnouncement(req.params.id, req.body);
   return successResponse(res, { data: result });
 });
 
 export const deleteAnnouncement = asyncHandler(async (req, res) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return errorResponse(res, {
+      statusCode: 400,
+      message: errors.array()[0].msg,
+    });
+  }
+
   await cmsService.deleteAnnouncement(req.params.id);
   return successResponse(res, { message: "Announcement deleted" });
 });
